Moved DeviceView and DeviceCreate* to their own folders
