+++ app/imports/ui/target-types/device/device-create/DeviceCreate.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { Meteor } from 'meteor/meteor';
 import { withTracker } from 'meteor/react-meteor-data';
 
-import Device from '../../../api/device/device';
+import Device from '../../../../api/device/device';
 import { DeviceCreateForm } from './DeviceCreateForm';
 
 export default DeviceCreate = withTracker( (props) => {
