+++ app/imports/ui/target-types/device/device-view/DeviceView.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
 import { Meteor } from 'meteor/meteor';
 import { withTracker } from 'meteor/react-meteor-data';
+import { Device } from '../../../../api/device/device';
 
-import { Device } from '../../../api/device/device';
+import styles from './DeviceView.module.scss';
 
 class DeviceView extends Component {
